squashed impl and runp projects
