+++ impl/src/test/scala/quasar/CountingRegressionSpec.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package quasar.run
+package quasar.impl
 
 import slamdata.Predef.{Map => SMap, _}
 import quasar._
