+++ impl/src/test/scala/quasar/QScriptRegressionSpec.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package quasar.run
+package quasar.impl
 
 import slamdata.Predef._
