--- run/src/main/scala/quasar/run/package.scala
@@ -1,24 +1,0 @@
-/*
- * Copyright 2020 Precog Data
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package quasar
-
-import quasar.contrib.scalaz.MonadError_
-
-package object run {
-  type MonadQuasarErr[F[_]] = MonadError_[F, QuasarError]
-  def MonadQuasarErr[F[_]](implicit ev: MonadQuasarErr[F]): MonadQuasarErr[F] = ev
-}
